requirements.txt
streamlit
easyocr
pandas
pillow
numpy